COUR-1961: feature: implement build_database_url function and refactor usage
