+++ Databases/PostgreSQL/orms/diesel-tutorial/excersices/src/lib.rs
@@ -1,0 +1,9 @@
+pub fn build_database_url(
+    host: &str,
+    port: u16,
+    database: &str,
+    user: &str,
+    password: &str,
+) -> String {
+    format!("postgres://{user}:{password}@{host}:{port}/{database}")
+}
